+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Changelog for NeoFS Node
 - `neofs-adm fschain load-report` command (#3649)
 - SN now supports new `getInfo` and `createV2` methods of the Container contract (#3670)
 - IR now supports container creation requests submitted via new `createV2` contract method (#3670)
+- IR structures containers in the contract iteratively (#3670)
 
 ### Fixed
 - Write cache using too much CPU (#3642)